2to3

+++ kegbot/pycore/flow_meter.py
@@ -45,7 +45,7 @@ class FlowMeter(object):
     The value `ticks` is always saved as `_last_ticks` for use in the next
     report.
     """
-    ticks = long(ticks)
+    ticks = int(ticks)
     self._logger.info('SetTicks: ticks=%s last=%s total=%s' % (
         ticks, self._last_ticks, self._total_ticks))
